Add database connection, modify register and login functionality to use database, modify section repository to use database. And add jBCrypt for hash user password.

+++ src/commands/RatePostCommand.java
@@ -35,7 +35,7 @@ public class RatePostCommand extends Command {
 		}
 		
 		int postId = Integer.parseInt(this.getData().get(0));
-		int grade = Integer.parseInt(this.getData().get(1));
+		byte grade = Byte.parseByte(this.getData().get(1));
 		
 		PostRepository.getInstance().addGradeToPost(postId, grade);
